Use size_t in for loops setting color palettes.

+++ src/loaders/ansi.c
@@ -491,14 +491,14 @@ ansilove_ansi(struct ansilove_ctx *ctx, struct ansilove_options *options)
 	} else if (workbench) {
 		gdImageFill(canvas, 0, 0, 0);
 
-		for (int i = 0; i < 16; i++)
+		for (size_t i = 0; i < 16; i++)
 			colors[i] = gdImageColorAllocate(canvas, workbench_palette[i*3],
 			    workbench_palette[i*3+1],
 			    workbench_palette[i*3+2]);
 	} else {
 		/* Allocate standard ANSi color palette */
 
-		for (int i = 0; i < 16; i++)
+		for (size_t i = 0; i < 16; i++)
 			colors[i] = gdImageColorAllocate(canvas, ansi_palette[i*3],
 			    ansi_palette[i*3+1],
 			    ansi_palette[i*3+2]);
