+++ src/loaders/pcboard.c
@@ -164,7 +164,7 @@ ansilove_pcboard(struct ansilove_ctx *ctx, struct ansilove_options *options)
 	/* allocate color palette */
 	uint32_t colors[16];
 
-	for (int i = 0; i < 16; i++) {
+	for (size_t i = 0; i < 16; i++) {
 		colors[i] = gdImageColorAllocate(canvas, ansi_palette[i*3],
 		    ansi_palette[i*3+1],
 		    ansi_palette[i*3+2]);
